Create Code
Consider the words are NOT case sensitive (which means
"Jingle" and "jingle" are the same word). Calculate word count
(occurrence of each word) in the file alice.txt and show the top 200 words with their count in descending order.
Sample output:
Word Count
jingle 7
hello 4

Consider the words are NOT case sensitive (which means
"Jingle" and "jingle" are the same word). Show the average length
of words starting with each letter.
Sample output:
a 2.5
b 5
c 3.2

Consider the words are NOT case sensitive (which means
"Jingle" and "jingle" are the same word). Show top 10 starting letters having highest average length words. Also show the total word count for these starting letters.
Sample output:
<starting letter> <word count> <avg word length>
b                                  17                   5
c                                  11                   3.2
a                                  20                   2.5